persist: move impls of admin and inspect cli into src/persist-client
Two immediate motivations:

- We originally had them outside of src/persist-client to avoid the clap
  and serde_json deps, but now we get both of them via workspace-hack,
  the math on this has changed. Having them inside src/persist-client
  itself means we don't have to cross a crate boundary, which will make
  it lower cost to add more of these. They've already come in incredibly
  handy for debugging that it makes sense to optimize for making it easy
  to add more (and to maintain the ones we have).
- I have vague aspirations of creating a persistcli wrapper in the cloud
  repo that has special cases for our various prod values of
  consensus_uri and blob_uri baked in, for ease of use.
