+++ src/persist-client/Cargo.toml
@@ -33,6 +33,7 @@ anyhow = { version = "1.0.66", features = ["backtrace"] }
 async-stream = "0.3.3"
 async-trait = "0.1.59"
 bytes = "1.3.0"
+clap = { version = "3.2.20", features = [ "derive" ] }
 differential-dataflow = { git = "https://github.com/TimelyDataflow/differential-dataflow.git" }
 futures = "0.3.25"
 futures-util = "0.3"
@@ -50,6 +51,7 @@ prost = { version = "0.11.3", features = ["no-recursion-limit"] }
 sentry-tracing = "0.29.1"
 semver = "1.0.16"
 serde = { version = "1.0.147", features = ["derive"] }
+serde_json = "1.0.89"
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", default-features = false, features = ["bincode"] }
 tokio = { version = "1.23.0", default-features = false, features = ["macros", "sync", "rt", "rt-multi-thread", "time"] }
 tracing = "0.1.37"
