+++ src/persist-client/src/cli/admin.rs
@@ -9,22 +9,110 @@
 
 //! CLI introspection tools for persist
 
+use std::str::FromStr;
 use std::sync::Arc;
 use std::time::Instant;
 
 use anyhow::anyhow;
 use mz_ore::metrics::MetricsRegistry;
+use mz_ore::now::SYSTEM_TIME;
 use mz_persist::cfg::{BlobConfig, ConsensusConfig};
 use mz_persist::location::{Blob, Consensus};
+use prometheus::proto::{MetricFamily, MetricType};
 use tracing::info;
 
 use crate::async_runtime::CpuHeavyRuntime;
+use crate::cli::inspect::StateArgs;
 use crate::internal::compact::{CompactReq, Compactor};
 use crate::internal::machine::Machine;
 use crate::internal::metrics::{MetricsBlob, MetricsConsensus};
 use crate::internal::trace::{ApplyMergeResult, FueledMergeRes};
 use crate::write::WriterId;
-use crate::{Metrics, PersistConfig, ShardId, StateVersions};
+use crate::{Metrics, PersistConfig, ShardId, StateVersions, BUILD_INFO};
+
+/// Commands for read-write administration of persist state
+#[derive(Debug, clap::Args)]
+pub struct AdminArgs {
+    #[clap(subcommand)]
+    command: Command,
+
+    /// Whether to commit any modifications (defaults to dry run).
+    #[clap(long)]
+    pub(crate) commit: bool,
+}
+
+/// Individual subcommands of admin
+#[derive(Debug, clap::Subcommand)]
+pub(crate) enum Command {
+    /// Manually completes all fueled compactions in a shard.
+    ForceCompaction(ForceCompactionArgs),
+}
+
+/// Manually completes all fueled compactions in a shard.
+#[derive(Debug, clap::Parser)]
+pub(crate) struct ForceCompactionArgs {
+    #[clap(flatten)]
+    state: StateArgs,
+
+    /// An upper bound on compaction's memory consumption.
+    #[clap(long, default_value_t = 0)]
+    compaction_memory_bound_bytes: usize,
+}
+
+/// Runs the given read-write admin command.
+pub async fn run(command: AdminArgs) -> Result<(), anyhow::Error> {
+    match command.command {
+        Command::ForceCompaction(args) => {
+            let shard_id = ShardId::from_str(&args.state.shard_id).expect("invalid shard id");
+            let mut cfg = PersistConfig::new(&BUILD_INFO, SYSTEM_TIME.clone());
+            if args.compaction_memory_bound_bytes > 0 {
+                cfg.compaction_memory_bound_bytes = args.compaction_memory_bound_bytes;
+            }
+            let metrics_registry = MetricsRegistry::new();
+            let () = force_compaction(
+                cfg,
+                &metrics_registry,
+                shard_id,
+                &args.state.consensus_uri,
+                &args.state.blob_uri,
+                command.commit,
+            )
+            .await?;
+            info_log_non_zero_metrics(&metrics_registry.gather());
+        }
+    }
+    Ok(())
+}
+
+fn info_log_non_zero_metrics(metric_families: &[MetricFamily]) {
+    for mf in metric_families {
+        for m in mf.get_metric() {
+            let val = match mf.get_field_type() {
+                MetricType::COUNTER => m.get_counter().get_value(),
+                MetricType::GAUGE => m.get_gauge().get_value(),
+                x => unimplemented!("unhandled metric type: {:?}", x),
+            };
+            if val == 0.0 {
+                continue;
+            }
+            let label_pairs = m.get_label();
+            let mut labels = String::new();
+            if !label_pairs.is_empty() {
+                labels.push_str("{");
+                for lb in label_pairs {
+                    if labels != "{" {
+                        labels.push_str(",");
+                    }
+                    labels.push_str(lb.get_name());
+                    labels.push_str(":");
+                    labels.push_str(lb.get_value());
+                }
+                labels.push_str("}");
+            }
+            info!("{}{} {}", mf.get_name(), labels, val);
+        }
+    }
+}
 
 /// Manually completes all fueled compactions in a shard.
 pub async fn force_compaction(
@@ -61,7 +149,7 @@ pub async fn force_compaction(
         .await;
     loop {
         let state_res = state_versions
-            .fetch_current_state::<crate::inspect::K, crate::inspect::V, u64, i64>(
+            .fetch_current_state::<crate::cli::inspect::K, crate::cli::inspect::V, u64, i64>(
                 &shard_id,
                 versions.0.clone(),
             )
@@ -69,7 +157,7 @@ pub async fn force_compaction(
         let state = match state_res {
             Ok(state) => state,
             Err(codec) => {
-                let mut kvtd = crate::inspect::KVTD_CODECS.lock().expect("lockable");
+                let mut kvtd = crate::cli::inspect::KVTD_CODECS.lock().expect("lockable");
                 *kvtd = codec.actual;
                 continue;
             }
@@ -84,7 +172,7 @@ pub async fn force_compaction(
         break;
     }
 
-    let mut machine = Machine::<crate::inspect::K, crate::inspect::V, u64, i64>::new(
+    let mut machine = Machine::<crate::cli::inspect::K, crate::cli::inspect::V, u64, i64>::new(
         cfg.clone(),
         shard_id,
         Arc::clone(&metrics),
@@ -138,15 +226,16 @@ pub async fn force_compaction(
                 info!("skipping compaction because --commit is not set");
                 continue;
             }
-            let res = Compactor::<crate::inspect::K, crate::inspect::V, u64, i64>::compact(
-                cfg.clone(),
-                Arc::clone(&blob),
-                Arc::clone(&metrics),
-                Arc::new(CpuHeavyRuntime::new()),
-                req,
-                writer_id.clone(),
-            )
-            .await?;
+            let res =
+                Compactor::<crate::cli::inspect::K, crate::cli::inspect::V, u64, i64>::compact(
+                    cfg.clone(),
+                    Arc::clone(&blob),
+                    Arc::clone(&metrics),
+                    Arc::new(CpuHeavyRuntime::new()),
+                    req,
+                    writer_id.clone(),
+                )
+                .await?;
             info!(
                 "attempt {} req {}: compacted into {} parts {} bytes in {:?}",
                 attempt,
