+++ src/persist-client/src/cli/inspect.rs
@@ -11,6 +11,7 @@
 
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::fmt;
+use std::str::FromStr;
 use std::sync::{Arc, Mutex};
 
 use anyhow::anyhow;
@@ -27,6 +28,7 @@ use mz_persist::cfg::{BlobConfig, ConsensusConfig};
 use mz_persist::indexed::encoding::BlobTraceBatchPart;
 use mz_persist_types::{Codec, Codec64};
 use mz_proto::RustType;
+use serde_json::json;
 
 use crate::fetch::EncodedPart;
 use crate::internal::paths::{
@@ -40,6 +42,156 @@ const READ_ALL_BUILD_INFO: BuildInfo = BuildInfo {
     sha: "0000000000000000000000000000000000000000",
     time: "",
 };
+
+/// Commands for read-only inspection of persist state
+#[derive(Debug, clap::Args)]
+pub struct InspectArgs {
+    #[clap(subcommand)]
+    command: Command,
+}
+
+/// Individual subcommands of inspect
+#[derive(Debug, clap::Subcommand)]
+pub(crate) enum Command {
+    /// Prints latest consensus state as JSON
+    State(StateArgs),
+
+    /// Prints latest consensus rollup state as JSON
+    StateRollup(StateArgs),
+
+    /// Prints consensus rollup state of all known rollups as JSON
+    StateRollups(StateArgs),
+
+    /// Prints the count and size of blobs in an environment
+    BlobCount(BlobCountArgs),
+
+    /// Prints blob batch part contents
+    BlobBatchPart(BlobBatchPartArgs),
+
+    /// Prints the unreferenced blobs across all shards
+    UnreferencedBlobs(StateArgs),
+
+    /// Prints each consensus state change as JSON. Output includes the full consensus state
+    /// before and after each state transitions:
+    ///
+    /// ```text
+    /// {
+    ///     "previous": previous_consensus_state,
+    ///     "new": new_consensus_state,
+    /// }
+    /// ```
+    ///
+    /// This is most helpfully consumed using a JSON diff tool like `jd`. A useful incantation
+    /// to show only the changed fields between state transitions:
+    ///
+    /// ```text
+    /// persistcli inspect state-diff --shard-id <shard> --consensus-uri <consensus_uri> |
+    ///     while read diff; do
+    ///         echo $diff | jq '.new' > temp_new
+    ///         echo $diff | jq '.previous' > temp_previous
+    ///         echo $diff | jq '.new.seqno'
+    ///         jd -color -set temp_previous temp_new
+    ///     done
+    /// ```
+    ///
+    #[clap(verbatim_doc_comment)]
+    StateDiff(StateArgs),
+}
+
+/// Runs the given read-only inspect command.
+pub async fn run(command: InspectArgs) -> Result<(), anyhow::Error> {
+    match command.command {
+        Command::State(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let state = fetch_latest_state(shard_id, &args.consensus_uri, &args.blob_uri).await?;
+            println!(
+                "{}",
+                serde_json::to_string_pretty(&state).expect("unserializable state")
+            );
+        }
+        Command::StateRollup(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let state_rollup =
+                fetch_latest_state_rollup(shard_id, &args.consensus_uri, &args.blob_uri).await?;
+            println!(
+                "{}",
+                serde_json::to_string_pretty(&state_rollup).expect("unserializable state")
+            );
+        }
+        Command::StateRollups(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let state_rollups =
+                fetch_state_rollups(shard_id, &args.consensus_uri, &args.blob_uri).await?;
+            println!(
+                "{}",
+                serde_json::to_string_pretty(&state_rollups).expect("unserializable state")
+            );
+        }
+        Command::StateDiff(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let states = fetch_state_diffs(shard_id, &args.consensus_uri, &args.blob_uri).await?;
+            for window in states.windows(2) {
+                println!(
+                    "{}",
+                    json!({
+                        "previous": window[0],
+                        "new": window[1]
+                    })
+                );
+            }
+        }
+        Command::BlobCount(args) => {
+            let blob_counts = blob_counts(&args.blob_uri).await?;
+            println!("{}", json!(blob_counts));
+        }
+        Command::BlobBatchPart(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let updates = blob_batch_part(&args.blob_uri, shard_id, args.key, args.limit).await?;
+            println!("{}", json!(updates));
+        }
+        Command::UnreferencedBlobs(args) => {
+            let shard_id = ShardId::from_str(&args.shard_id).expect("invalid shard id");
+            let unreferenced_blobs =
+                unreferenced_blobs(&shard_id, &args.consensus_uri, &args.blob_uri).await?;
+            println!("{}", json!(unreferenced_blobs));
+        }
+    }
+
+    Ok(())
+}
+
+/// Arguments for viewing the current state of a given shard
+#[derive(Debug, Clone, clap::Parser)]
+pub struct StateArgs {
+    /// Shard to view
+    #[clap(long)]
+    pub(crate) shard_id: String,
+
+    /// Consensus to use.
+    ///
+    /// When connecting to a deployed environment's consensus table, the Postgres/CRDB connection
+    /// string must contain the database name and `options=--search_path=consensus`.
+    ///
+    /// When connecting to Cockroach Cloud, use the following format:
+    ///
+    /// ```text
+    /// postgresql://<user>:$COCKROACH_PW@<hostname>:<port>/environment_<environment-id>
+    ///   ?sslmode=verify-full
+    ///   &sslrootcert=/path/to/cockroach-cloud/certs/cluster-ca.crt
+    ///   &options=--search_path=consensus
+    /// ```
+    ///
+    #[clap(long, verbatim_doc_comment, env = "CONSENSUS_URI")]
+    pub(crate) consensus_uri: String,
+
+    /// Blob to use
+    ///
+    /// When connecting to a deployed environment's blob, the necessary connection glue must be in
+    /// place. e.g. for S3, sign into SSO, set AWS_PROFILE and AWS_REGION appropriately, with a blob
+    /// URI scoped to the environment's bucket prefix.
+    #[clap(long, env = "BLOB_URI")]
+    pub(crate) blob_uri: String,
+}
 
 /// Fetches the current state of a given shard
 pub async fn fetch_latest_state(
@@ -219,6 +371,30 @@ pub async fn fetch_state_diffs(
     Ok(live_states)
 }
 
+/// Arguments for viewing contents of a batch part
+#[derive(Debug, Clone, clap::Parser)]
+pub struct BlobBatchPartArgs {
+    /// Shard to view
+    #[clap(long)]
+    shard_id: String,
+
+    /// Blob key (without shard)
+    #[clap(long)]
+    key: String,
+
+    /// Blob to use
+    ///
+    /// When connecting to a deployed environment's blob, the necessary connection glue must be in
+    /// place. e.g. for S3, sign into SSO, set AWS_PROFILE and AWS_REGION appropriately, with a blob
+    /// URI scoped to the environment's bucket prefix.
+    #[clap(long)]
+    blob_uri: String,
+
+    /// Number of updates to output. Default is unbounded.
+    #[clap(long, default_value = "18446744073709551615")]
+    limit: usize,
+}
+
 #[derive(Debug, serde::Serialize)]
 struct BatchPartOutput {
     desc: Description<u64>,
@@ -282,6 +458,18 @@ pub async fn blob_batch_part(
     }
 
     Ok(out)
+}
+
+/// Arguments for viewing the blobs of a given shard
+#[derive(Debug, Clone, clap::Parser)]
+pub struct BlobCountArgs {
+    /// Blob to use
+    ///
+    /// When connecting to a deployed environment's blob, the necessary connection glue must be in
+    /// place. e.g. for S3, sign into SSO, set AWS_PROFILE and AWS_REGION appropriately, with a blob
+    /// URI scoped to the environment's bucket prefix.
+    #[clap(long)]
+    blob_uri: String,
 }
 
 #[derive(Debug, Default, serde::Serialize)]
